fix(31-07): B-GAP-3 time_t width audit (GroupObject + SwgCuiGroup) + Rule-3 callback fix
time_t per-site verdicts (B-GAP-3):
- GroupObject.cpp:413/424 -> DISPLAY. The std::pair<time_t,time_t> epoch MEMBER
  stays full-width (live client-epoch state, never serialized raw). Both returns
  compute a seconds-DURATION for a UI countdown (SwgCuiGroup/SwgCuiStatusGround);
  small value, explicit static_cast<unsigned int> (x64 C4244-clean).
- SwgCuiGroup.cpp:269/272 -> DISPLAY. timerEnd/timerTotal retyped time_t ->
  unsigned int to match the getter return contract (they are seconds-durations,
  not timestamps); removes the time_t->long/unsigned int narrowings at the root
  (pixel-width math + CalendarTime::convertSecondsToMS(unsigned int)).

Rule-3 blocking fix (same TU): GroupObject members/memberShips OnInsert/OnErase
callbacks const size_t -> const unsigned int to match AutoDeltaVector's wire-
stable const-unsigned-int callback contract (size_t==unsigned int only on 32-bit;
on x64 the pointer-to-member type mismatched -> C2664 blocked the TU). Used ONLY
as AutoDeltaVector callbacks.

Both target TUs compile x64 exit 0.

DEF-31-07: 2 unrelated std::distance->int C4244 (D-07/N2 class) logged, NOT fixed.
DEF-31-07-AUTODELTAMAP: NEW pre-existing class-(B) escape (AutoDeltaMap.h
Archive put/get(size_t) C2668/C2665, same defect class as the 31-05 std::map fix)
flagged STOP-and-report -- multi-site serialization-width change in a shared
external/ours template header, exceeds this plan's 4-task scope. NOT fixed here.

+++ src/engine/client/library/clientGame/src/shared/object/GroupObject.cpp
@@ -125,7 +125,7 @@ void GroupObject::membersOnChanged   ()
 
 //----------------------------------------------------------------------
 
-void GroupObject::membersOnInsert   (const size_t, const GroupMember & member)
+void GroupObject::membersOnInsert   (const unsigned int, const GroupMember & member)
 {
 	Transceivers::memberAdded.emitMessage (Messages::MemberAdded::Payload (this, &member));
 
@@ -140,7 +140,7 @@ void GroupObject::membersOnInsert   (const size_t, const GroupMember & member)
 
 //----------------------------------------------------------------------
 
-void GroupObject::membersOnErase  (const size_t, const GroupMember & member)
+void GroupObject::membersOnErase  (const unsigned int, const GroupMember & member)
 {
 	Transceivers::memberRemoved.emitMessage (Messages::MemberRemoved::Payload (this, &member));
 
@@ -162,14 +162,14 @@ void GroupObject::memberShipsOnChanged()
 
 //----------------------------------------------------------------------
 
-void GroupObject::memberShipsOnInsert(const size_t, const GroupShipFormationMember & member)
+void GroupObject::memberShipsOnInsert(const unsigned int, const GroupShipFormationMember & member)
 {
 	Transceivers::memberShipAdded.emitMessage (Messages::MemberShipAdded::Payload (this, &member));
 }
 
 //----------------------------------------------------------------------
 
-void GroupObject::memberShipsOnErase(const size_t, const GroupShipFormationMember & member)
+void GroupObject::memberShipsOnErase(const unsigned int, const GroupShipFormationMember & member)
 {
 	Transceivers::memberShipRemoved.emitMessage (Messages::MemberShipRemoved::Payload (this, &member));
 }
@@ -410,7 +410,13 @@ unsigned int GroupObject::getSecondsLeftOnGroupPickup() const
 	{
 		time_t const timeNow = ::time(NULL);
 		if (m_groupPickupTimerClientEpoch.second > timeNow)
-			return (m_groupPickupTimerClientEpoch.second - timeNow);
+			// BITS-03 B-GAP-3 verdict: DISPLAY. The epoch pair member stays
+			// full-width time_t (live client-epoch state, never serialized raw);
+			// this returns a seconds-remaining DURATION for a UI countdown label
+			// (SwgCuiGroup/SwgCuiStatusGround). The value is small (a group
+			// pickup timer) and fits unsigned int on both bitnesses -- explicit
+			// static_cast makes the narrowing intentional (silences x64 C4244).
+			return static_cast<unsigned int>(m_groupPickupTimerClientEpoch.second - timeNow);
 	}
 
 	return 0;
@@ -421,7 +427,9 @@ unsigned int GroupObject::getSecondsLeftOnGroupPickup() const
 unsigned int GroupObject::getGroupPickupDurationSeconds() const
 {
 	if ((m_groupPickupTimerClientEpoch.first > 0) && (m_groupPickupTimerClientEpoch.second > 0) && (m_groupPickupTimerClientEpoch.second > m_groupPickupTimerClientEpoch.first))
-		return (m_groupPickupTimerClientEpoch.second - m_groupPickupTimerClientEpoch.first);
+		// BITS-03 B-GAP-3 verdict: DISPLAY. Total-duration seconds for the same
+		// UI countdown; small value, explicit static_cast (x64 C4244-clean).
+		return static_cast<unsigned int>(m_groupPickupTimerClientEpoch.second - m_groupPickupTimerClientEpoch.first);
 
 	return 0;
 }
